requirements.txt
fastapi
uvicorn
pydantic
pymupdf
pdfplumber
python-dotenv
pandas
jupyterlab
httpx
streamlit
